move positioning useEffect to provider
